+++ packages/react/src/components/Tree/Tree.tsx
@@ -20,6 +20,7 @@ import {
   WithAsProp,
   withSafeTypeForAs,
   ShorthandCollection,
+  ComponentEventHandler,
 } from '../../types'
 import { Accessibility } from '../../lib/accessibility/types'
 import { treeBehavior } from '../../lib/accessibility'
@@ -52,6 +53,13 @@ export interface TreeProps extends UIComponentProps, ChildrenComponentProps {
    * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
    */
   renderItemTitle?: ShorthandRenderFunction
+
+  /** Called when activeIndex changes.
+   *
+   * @param {SyntheticEvent} event - React's original SyntheticEvent.
+   * @param {object} data - All props and proposed value.
+   */
+  onActiveIndexChange?: ComponentEventHandler<TreeProps>
 }
 
 export interface TreeState {
@@ -85,6 +93,7 @@ class Tree extends AutoControlledComponent<WithAsProp<TreeProps>, TreeState> {
     items: customPropTypes.collectionShorthand,
     renderItemTitle: PropTypes.func,
     rtlAttributes: PropTypes.func,
+    onActiveIndexChange: PropTypes.func,
   }
 
   static defaultProps = {
@@ -111,8 +120,13 @@ class Tree extends AutoControlledComponent<WithAsProp<TreeProps>, TreeState> {
             return acc
           }, [])
         : []
-      this.trySetState({ activeIndex })
+      this.trySetActiveIndexAndTriggerEvent(e, activeIndex)
     },
+  }
+
+  trySetActiveIndexAndTriggerEvent = (e, activeIndex) => {
+    this.trySetState({ activeIndex })
+    _.invoke(this.props, 'onActiveIndexChange', e, { ...this.props, activeIndex })
   }
 
   getInitialAutoControlledState({ exclusive }): TreeState {
@@ -144,7 +158,7 @@ class Tree extends AutoControlledComponent<WithAsProp<TreeProps>, TreeState> {
 
   handleTreeItemOverrides = (predefinedProps: TreeItemProps) => ({
     onTitleClick: (e: React.SyntheticEvent, treeItemProps: TreeItemProps) => {
-      this.trySetState({ activeIndex: this.computeNewIndex(treeItemProps) })
+      this.trySetActiveIndexAndTriggerEvent(e, this.computeNewIndex(treeItemProps))
       _.invoke(predefinedProps, 'onTitleClick', e, treeItemProps)
     },
   })